Move configuration options out of catalog context and inside the service collection
