--- src/Smart.FA.Catalog.Infrastructure/Options/DALOptions.cs
@@ -1,6 +1,0 @@
-﻿namespace Infrastructure.Options;
-
-public class DALOptions
-{
-    public bool UseConsoleLogger { get; set; }
-}
