+++ src/Smart.FA.Catalog.Infrastructure/Options/EFCore.cs
@@ -1,0 +1,6 @@
+﻿namespace Smart.FA.Catalog.Infrastructure.Options;
+
+public class EFCore
+{
+    public bool UseConsoleLogger { get; set; }
+}
