+++ src/Smart.FA.Catalog.Web/appsettings.Development.json
@@ -24,7 +24,7 @@
   "MediatR": {
     "LogRequests" : true
   },
-  "DALOptions":
+  "EFCore":
   {
     "UseConsoleLogger": true
   }
