+++ src/Smart.FA.Catalog.Web/appsettings.Staging.json
@@ -21,7 +21,7 @@
       "NumberOfTrainingsDisplayed": 3
     }
   },
-  "DALOptions":
+  "EFCore":
   {
     "UseConsoleLogger": false
   }
