+++ src/Smart.FA.Catalog.Web/appsettings.json
@@ -21,9 +21,9 @@
       "NumberOfTrainingsDisplayed": 3
     }
   },
-  "DALOptions":
+  "EFCore":
   {
-    "UseConsoleLogger": false
+    "UseConsoleLogger": true
   },
   "MediatR": {
     "LogRequests": false
